feat: add custom code block editor and toolbar for enhanced code editing experience
- Introduced PlainTextCodeEditorDescriptor to manage code block behavior, including handling code block removal.
- Implemented CustomToolbar component with an Insert Code Block button for improved code insertion functionality.
- Updated InitializedMDXEditor to integrate the new custom toolbar and code editor, enhancing the overall markdown editing experience.
- Ensured accessibility features are included in both the toolbar and editor components, improving user interaction.
